Separate explorer IPC from host tools
Explorer IPC (electron/explorerIPC.ts) is now independent from agent
host tools (electron/hostTools.ts):
- explorer:list-directory, explorer:read-file, explorer:write-file
  — unsandboxed, for user's file explorer UI
- host-tools:call-tool (host_read/write/edit/search/bash)
  — sandboxed to per-agent allowed_paths, for agent use

Preload exposes separate window.electron.explorer API.
