+++ electron/hostTools.ts
@@ -454,87 +454,6 @@ export function registerHostToolIPC(): void {
     return HOST_TOOL_NAMES.has(name);
   });
 
-  // --- Explorer IPC (unsandboxed — for user's file explorer, not agent tools) ---
-
-  ipcMain.handle('host-tools:list-directory', (_event, dirPath: string) => {
-    // Root: return system drives / home directory
-    if (!dirPath) {
-      const home = process.env.HOME || process.env.USERPROFILE || '';
-      const roots: Array<{ name: string; fullPath: string; type: 'directory'; size: number; modified: string | null; extension: string }> = [];
-      if (process.platform === 'win32') {
-        // List common drive letters
-        for (const letter of ['C', 'D', 'E', 'F']) {
-          const drive = `${letter}:\\`;
-          if (fs.existsSync(drive)) {
-            roots.push({ name: `${letter}:`, fullPath: drive, type: 'directory', size: 0, modified: null, extension: '' });
-          }
-        }
-      } else {
-        roots.push({ name: 'Home', fullPath: home, type: 'directory', size: 0, modified: null, extension: '' });
-        roots.push({ name: '/', fullPath: '/', type: 'directory', size: 0, modified: null, extension: '' });
-      }
-      return { path: '', entries: roots, isRoot: true };
-    }
-
-    try {
-      const resolved = path.resolve(dirPath);
-      const stat = fs.statSync(resolved);
-      if (!stat.isDirectory()) {
-        return { path: resolved, entries: [], isRoot: false, error: 'Not a directory' };
-      }
-
-      const dirEntries = fs.readdirSync(resolved, { withFileTypes: true });
-      const entries = dirEntries
-        .filter(e => !e.name.startsWith('.'))
-        .map(entry => {
-          const fullPath = path.join(resolved, entry.name);
-          let size = 0;
-          let modified: string | null = null;
-          try {
-            const s = fs.statSync(fullPath);
-            size = s.size;
-            modified = s.mtime.toISOString();
-          } catch {}
-          return {
-            name: entry.name,
-            fullPath,
-            type: (entry.isDirectory() ? 'directory' : 'file') as 'directory' | 'file',
-            size,
-            modified,
-            extension: entry.isFile() ? path.extname(entry.name) : '',
-          };
-        })
-        .sort((a, b) => {
-          if (a.type !== b.type) return a.type === 'directory' ? -1 : 1;
-          return a.name.localeCompare(b.name);
-        });
-
-      return { path: resolved, entries, isRoot: false };
-    } catch (e: any) {
-      return { path: dirPath, entries: [], isRoot: false, error: e.message };
-    }
-  });
-
-  ipcMain.handle('host-tools:read-file', async (_event, filePath: string) => {
-    try {
-      const resolved = path.resolve(filePath);
-      const content = fs.readFileSync(resolved, { encoding: 'utf-8' });
-      return { content, path: resolved };
-    } catch (e: any) {
-      return { error: e.message };
-    }
-  });
-
-  ipcMain.handle('host-tools:write-file', async (_event, filePath: string, content: string) => {
-    try {
-      const resolved = path.resolve(filePath);
-      fs.mkdirSync(path.dirname(resolved), { recursive: true });
-      fs.writeFileSync(resolved, content, { encoding: 'utf-8' });
-      return { status: 'ok', path: resolved };
-    } catch (e: any) {
-      return { error: e.message };
-    }
-  });
 }
 
 export { HOST_TOOL_NAMES, getHostToolSchemas };
